Make Observable inspire 10 times.
In case of Observable, we create an infinite sequence which would keep
sending inspirations infinitely. In the dispose method of the Observable
we clear the interval. Then we declaratively take only 10 inspirations
from the sequence, and Observalbe gets rid of the interval for us.

In case of Promises, we need to be explicit about how and when the
interval is cleared. However for Observable, we just tell it to clear
the interval when Observable finishes, and then we use `take` operator
to complete the infinite Observable after getting 10 items.